Park func_8004733C on which argument gets which saved register
A two-way dispatch on the top bit of a card id. Body, guard, the unsigned
range test and the -0x7000 adjustment all match; retail holds arg1 in $s0 and
arg0 in $s1 and gcc does the reverse, which moves ten instructions. Naming
either argument first, an early return instead of the else, and narrowing the
second parameter all leave it at ten or worse; the permuter's best is 150.

Two useful signature facts came out of it though, both now applied:
func_80049230 and func_80045208 take their second argument as s16 -- the
caller sign-extends before the call, which only happens when the prototype
says so, and func_80049230 still matches with the narrower parameter because
it re-extends internally either way.

+++ src/func_80049230.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-void func_80049230(s16 arg0, s32 arg1) {
+void func_80049230(s16 arg0, s16 arg1) {
     Cmd2C c;
 
     if (arg0 < 0) {
